requirements.txt
python-dotenv
tavily-python
requests
beautifulsoup4
yutori
boto3